% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{AddNotifications}
\alias{AddNotifications}
\title{Add notifications on the client browser using errors, warnings or messages
generated when executing an R expression}
\usage{
AddNotifications(res)
}
\arguments{
\item{res}{A list containing errors, warnings and messages which were
generated when executing an R expression. See \code{\link{CatchToList}} for
more details.}
}
\value{
A list containing the notification id for error, warning and message.
}
\description{
Add notifications on the client browser using errors, warnings or messages
generated when executing an R expression
}
\keyword{internal}
