% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureEstimation.R
\name{bootstrapSigExposures}
\alias{bootstrapSigExposures}
\title{bootstrapSigExposures Function
This function allows to obtain the bootstrap distribution of the signature exposures of a certain tumor sample}
\usage{
bootstrapSigExposures(
  m,
  P,
  R,
  mutation.count = NULL,
  decomposition.method = decomposeQP,
  ...
)
}
\arguments{
\item{m}{observed tumor profile vector for a patient/sample, 96 by 1. It can be mutation
counts, or mutation probabilities.}

\item{P}{signature profile matrix, 96 by N (N = # signatures, COSMIC: N=30).}

\item{R}{The number of bootstrap replicates.}

\item{mutation.count}{if m is a vector of counts, then mutation.count equals the summation of all the counts.
If m is probabilities, then mutation.count has to be specified.}

\item{decompostion.method}{which method is selected to get the optimal solution: decomposeQP or decomposeSA}
}
\description{
bootstrapSigExposures Function
This function allows to obtain the bootstrap distribution of the signature exposures of a certain tumor sample
}
\keyword{bootstrap}
\keyword{internal}
