% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_subsets.R
\name{CancerTypeToSigSubset}
\alias{CancerTypeToSigSubset}
\title{Get the subset of signatures for a specified cancer type from a specified tumor cohort}
\usage{
CancerTypeToSigSubset(
  cancer.type,
  tumor.cohort = "PCAWG",
  sig.type = "SBS96",
  region = "genome"
)
}
\arguments{
\item{cancer.type}{Cancer type of the tumor, e.g. "Biliary-AdenoCA".}

\item{tumor.cohort}{The cohort of tumors to get information from package PCAWG7. Can
be "PCAWG", "TCGA", "other.genome", "other.exome".}

\item{sig.type}{Type of the signature, e.g. "SBS96".}

\item{region}{A character string designating a genomic region;
see \code{\link{as.catalog}} and \code{\link{ICAMS}}.}
}
\value{
The set of signatures found present in the specified \code{cancer.type} from
the specified \code{tumor.cohort}.
}
\description{
Get the subset of signatures for a specified cancer type from a specified tumor cohort
}
