% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{CatchToList}
\alias{CatchToList}
\title{Catch errors, warnings and messages generated when executing an R expression.}
\usage{
CatchToList(expr)
}
\arguments{
\item{expr}{An R expression to execute.}
}
\value{
A list containing errors, warnings and messages which were generated
  when executing \code{expr}.
}
\description{
Catch errors, warnings and messages generated when executing an R expression.
}
\keyword{internal}
