% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{CreateReadMeFile}
\alias{CreateReadMeFile}
\title{Create a README file describing the zip archive generated from VCF files}
\usage{
CreateReadMeFile(vcf.names, zipfile.name, ref.genome, region)
}
\arguments{
\item{vcf.names}{Names of VCF files uploaded by the user.}

\item{zipfile.name}{Name of zip archive specified by the user.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link[ICAMS]{ICAMS}}.}

\item{region}{A character string designating a genomic region;
see \code{\link[ICAMS]{as.catalog}} and \code{\link[ICAMS]{ICAMS}}.}
}
\description{
Create a README file describing the zip archive generated from VCF files
}
\keyword{internal}
