% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureEstimation.R
\name{decomposeQP}
\alias{decomposeQP}
\title{decomposeQP Function}
\usage{
decomposeQP(m, P, ...)
}
\arguments{
\item{m}{observed tumor profile vector for a single patient/sample, 96 by 1. m is normalized.}

\item{P}{signature profile matrix, 96 by N(N = # signatures, COSMIC: N=30)}

\item{...}{some control parameter that can be passed into the solve.QP function}
}
\description{
This function allows to get the optimal solution by using dual method to solve the quadratic programming problem.
}
\keyword{programming}
\keyword{quadratic}
