% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureEstimation.R
\name{decomposeSA}
\alias{decomposeSA}
\title{decomposeSA Function}
\usage{
decomposeSA(m, P, control = list())
}
\arguments{
\item{m}{observed tumor profile vector for a single patient/sample, 96 by 1. m is normalized.}

\item{P}{signature profile matrix, 96 by N(N = # signatures, COSMIC: N=30)}

\item{control}{some control parameter that can be passed into the GenSA function}
}
\description{
This function allows to get the optimal solution by using simulated annealing to solve the optimization problem.
}
\keyword{annealing}
\keyword{simulated}
