% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureEstimation.R
\name{findSigExposures}
\alias{findSigExposures}
\title{findSigExposures Function
wrapper function
This function allows to obtain the optimal solution by specifying quadratic programming or simulated annealing to solve the optimization problem.}
\usage{
findSigExposures(M, P, decomposition.method = decomposeQP, ...)
}
\arguments{
\item{M}{observed tumor profile matrix for all the patient/sample, 96 by G. G is the number of patients. Each column can be mutation
counts, or mutation probabilities. Each column will be normalized to sum up to 1.}

\item{P}{signature profile matrix, 96 by N(N = # signatures, COSMIC: N=30)}

\item{decompostion.method}{which method is selected to get the optimal solution: decomposeQP or decomposeSA}
}
\description{
findSigExposures Function
wrapper function
This function allows to obtain the optimal solution by specifying quadratic programming or simulated annealing to solve the optimization problem.
}
\keyword{QP}
\keyword{SA}
\keyword{internal}
\keyword{method:}
\keyword{optimal}
\keyword{or}
