% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{GenerateZipFileFromStrelkaIDVCFs}
\alias{GenerateZipFileFromStrelkaIDVCFs}
\title{This function generates a zip archive from Strelka ID VCF files.}
\usage{
GenerateZipFileFromStrelkaIDVCFs(
  files,
  zipfile,
  vcf.names,
  zipfile.name,
  ref.genome,
  region = "unknown",
  names.of.VCFs = NULL,
  base.filename = "",
  updateProgress = NULL
)
}
\arguments{
\item{files}{Character vector of file paths to the Strelka ID VCF files.}

\item{zipfile}{Pathname of the zip file to be created.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link[ICAMS]{ICAMS}}.}

\item{region}{A character string designating a genomic region;
see \code{\link[ICAMS]{as.catalog}} and \code{\link[ICAMS]{ICAMS}}.}

\item{names.of.VCFs}{Optional. Character vector of names of the VCF files.
The order of names in \code{names.of.VCFs} should match the order of VCFs
listed in \code{dir}. If \code{NULL}(default), this function will remove
all of the path up to and including the last path separator (if any) in
\code{dir} and file paths without extensions (and the leading dot) will be
used as the names of the VCF files.}

\item{base.filename}{Optional. The base name of the CSV and PDF files to be
produced; multiple files will be generated, each ending in
\eqn{x}\code{.csv} or \eqn{x}\code{.pdf}, where \eqn{x} indicates the type
of catalog.}

\item{updateProgress}{Optional. Currently only used in ICAMS.shiny package to
update the progress indicator.}
}
\description{
This function generates a zip archive from Strelka ID VCF files.
}
\keyword{internal}
