% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{GenerateZipFileFromVCFs}
\alias{GenerateZipFileFromVCFs}
\title{This function generates a zip archive from VCFs}
\usage{
GenerateZipFileFromVCFs(
  files,
  zipfile,
  vcf.names,
  zipfile.name,
  ref.genome,
  variant.caller,
  mergeSBS,
  num.of.cores,
  trans.ranges = NULL,
  region = "unknown",
  names.of.VCFs = NULL,
  base.filename = "",
  updateProgress = NULL
)
}
\arguments{
\item{files}{Character vector of file paths to the VCF files.}

\item{zipfile}{Pathname of the zip file to be created.}

\item{vcf.names}{Names of VCF files uploaded by the user.}

\item{zipfile.name}{Name of zip archive specified by the user.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link[ICAMS]{ICAMS}}.}

\item{variant.caller}{Name of the variant caller that produces the VCF, can
be either \code{strelka}, \code{mutect} or \code{freebayes}. This
information is needed to calculate the VAFs (variant allele frequencies).
If \code{"unknown"}(default) and \code{get.vaf.function} is NULL, then VAF
and read depth will be NAs.}

\item{mergeSBS}{Whether to merge adjacent SBSs as DBS. Value can be "yes" or
"no".}

\item{num.of.cores}{The number of cores to use. Not available on Windows
unless \code{num.of.cores = 1}.}

\item{trans.ranges}{Optional. If \code{ref.genome} specifies one of the
\code{\link[BSgenome]{BSgenome}} object
\enumerate{
  \item \code{\link[BSgenome.Hsapiens.1000genomes.hs37d5]{BSgenome.Hsapiens.1000genomes.hs37d5}}
  \item \code{\link[BSgenome.Hsapiens.UCSC.hg38]{BSgenome.Hsapiens.UCSC.hg38}}
  \item \code{\link[BSgenome.Mmusculus.UCSC.mm10]{BSgenome.Mmusculus.UCSC.mm10}}
}
then the function will infer \code{trans.ranges} automatically. Otherwise,
user will need to provide the necessary \code{trans.ranges}. Please refer to
\code{\link[ICAMS]{TranscriptRanges}} for more details.
If \code{is.null(trans.ranges)} do not add transcript range
information.}

\item{region}{A character string designating a genomic region;
see \code{\link[ICAMS]{as.catalog}} and \code{\link[ICAMS]{ICAMS}}.}

\item{names.of.VCFs}{Optional. Character vector of names of the VCF files.
The order of names in \code{names.of.VCFs} should match the order of VCFs
listed in \code{files}. If \code{NULL}(default), this function will remove
all of the path up to and including the last path separator (if any) in
\code{files} and file paths without extensions (and the leading dot) will be
used as the names of the VCF files.}

\item{base.filename}{Optional. The base name of the CSV and PDF files to be
produced; multiple files will be generated, each ending in
\eqn{x}\code{.csv} or \eqn{x}\code{.pdf}, where \eqn{x} indicates the type
of catalog.}

\item{updateProgress}{A callback function to update the progress indicator on
the user interface.}
}
\description{
This function generates a zip archive from VCFs
}
\keyword{internal}
