% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_attribution.R
\name{GetExposureWithConfidence}
\alias{GetExposureWithConfidence}
\title{Get signature exposure for one sample with confidence intervals}
\usage{
GetExposureWithConfidence(
  catalog,
  sig.universe,
  num.of.bootstrap.replicates = 1000,
  method = decomposeQP,
  conf.int = 0.95
)
}
\arguments{
\item{catalog}{A 1-column \strong{counts} catalog as defined in \code{\link{ICAMS}}.}

\item{sig.universe}{The universe of signatures used to do signature attribution.}

\item{num.of.bootstrap.replicates}{The number of bootstrap replicates.}

\item{method}{Method to get the optimal solution for signature attribution.}

\item{conf.int}{The required confidence interval.}
}
\description{
Get signature exposure for one sample with confidence intervals
}
\section{Value}{

  A matrix with one row per signature, with each row showing
  lower and upper bounds of the confidence
  interval and the mean of the estimated exposure,
 only for those signatures with a lower bound  > 0.
}

