% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{GetNamesOfVCFs}
\alias{GetNamesOfVCFs}
\title{This function extracts the sample names representing different VCF files
specified by user.}
\usage{
GetNamesOfVCFs(names.of.VCFs)
}
\arguments{
\item{names.of.VCFs}{A character string containing the sample names
representing different VCF files (different names separated by comma)
specified by user on the browser.}
}
\value{
A character vetor containing the sample names representing different
  VCF files.
}
\description{
This function extracts the sample names representing different VCF files
specified by user.
}
\keyword{internal}
