% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{GetTumorColNames}
\alias{GetTumorColNames}
\title{This function extracts names of columns which contain the tumor sample
information in Mutect VCFs.}
\usage{
GetTumorColNames(tumor.col.names)
}
\arguments{
\item{tumor.col.names}{A character string containing names of columns which
contain the tumor sample information in Mutect VCFs (different names
separated by comma) specified by user on the browser.}
}
\value{
A character vector containing the name of column which contains the
  tumor sample information in Mutect VCFs.
}
\description{
This function extracts names of columns which contain the tumor sample
information in Mutect VCFs.
}
\keyword{internal}
