% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{ICAMS.shinyApp}
\alias{ICAMS.shinyApp}
\title{Run the Shiny Application}
\usage{
ICAMS.shinyApp(...)
}
\arguments{
\item{...}{Further arguments to be passed to the function.}
}
\description{
Run the Shiny Application
}
