% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{PlotListOfCatalogsToPdf}
\alias{PlotListOfCatalogsToPdf}
\title{Title}
\usage{
PlotListOfCatalogsToPdf(
  list.of.catalogs,
  file,
  plot.SBS12,
  cex = 0.8,
  grid = TRUE,
  upper = TRUE,
  xlabels = TRUE,
  ylim = NULL
)
}
\arguments{
\item{list.of.catalogs}{}

\item{file}{}

\item{plot.SBS12}{}

\item{cex}{}

\item{grid}{}

\item{upper}{}

\item{xlabels}{}

\item{ylim}{}
}
\value{

}
\description{
Title
}
