% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{PrepareExampleVCFs}
\alias{PrepareExampleVCFs}
\title{Prepare test VCFs for user to test}
\usage{
PrepareExampleVCFs(file)
}
\arguments{
\item{file}{Path of the file to be written.}
}
\description{
Prepare test VCFs for user to test
}
\keyword{internal}
