% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{PrepareSampleSpectra}
\alias{PrepareSampleSpectra}
\title{Prepare test catalogs for user to test}
\usage{
PrepareSampleSpectra(file)
}
\arguments{
\item{file}{Path of the file to be written.}
}
\description{
Prepare test catalogs for user to test
}
\keyword{internal}
