% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{ProcessMutectVCFs}
\alias{ProcessMutectVCFs}
\title{This function is a wrapper function processing Mutect VCF files to
generate a zip archive.}
\usage{
ProcessMutectVCFs(input, output, file, ids)
}
\arguments{
\item{input}{A list-like object used in shiny app to store the current values
of all of the widgets in the app.}

\item{output}{A list-like object used in shiny app that stores instructions
for building the R objects in the app.}

\item{file}{A file path (string) of a nonexistent temp file, using which the
function writes the content to that file path. See
\code{\link[shiny]{downloadHandler}} for more details.}

\item{ids}{A list containing the existing notification ids for error, warning
and message.}
}
\value{
A list of updated notification ids for error, warning and message
  after running this function.
}
\description{
This function is a wrapper function processing Mutect VCF files to
generate a zip archive.
}
\keyword{internal}
