% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{RemoveAllNotifications}
\alias{RemoveAllNotifications}
\title{Remove all notifications on the client browser.}
\usage{
RemoveAllNotifications(ids)
}
\arguments{
\item{ids}{A list of notification ids which are to be removed.}
}
\description{
Remove all notifications on the client browser.
}
\keyword{internal}
