% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureEstimation.R
\name{suboptimalSigExposures}
\alias{suboptimalSigExposures}
\title{suboptimalSigExposures Function
This function allows to obtain the simulated annealing distribution of the signature exposures of a certain tumor sample}
\usage{
suboptimalSigExposures(
  m,
  P,
  R,
  optimal.error = NULL,
  suboptimal.factor = 1.05,
  control = list()
)
}
\arguments{
\item{m}{observed tumor profile vector for a patient/sample, 96 by 1. It can be mutation
counts, or mutation probabilities.}

\item{P}{signature profile matrix, 96 by N(N = # signatures, COSMIC: N=30)}

\item{R}{The number of replicates/trials of simulated annealing.}

\item{optimal.error}{if it is NULL, then use SA method to obtain. Or you can provide one.}

\item{suboptimal.factor}{suboptimal error.}

\item{control}{some control parameter that can be passed into the function}

\item{mutation.count}{if m is a vector of counts, then mutation.count equals the summation of all the counts.
If m is probabilities, then mutation.count has to be specified.}
}
\description{
suboptimalSigExposures Function
This function allows to obtain the simulated annealing distribution of the signature exposures of a certain tumor sample
}
\keyword{'simulated}
\keyword{annealing'}
\keyword{internal}
\keyword{suboptimal}
