% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{TransCountsCatalogToDensity}
\alias{TransCountsCatalogToDensity}
\title{Transfrom a list of counts catalogs to a list of density catalogs}
\usage{
TransCountsCatalogToDensity(list)
}
\arguments{
\item{list}{A list of counts catalogs.}
}
\value{
A list of density catalogs transformed from \code{list}.
}
\description{
Transfrom a list of counts catalogs to a list of density catalogs
}
\keyword{internal}
