% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{UpdateNotificationIDs}
\alias{UpdateNotificationIDs}
\title{Update the existing list of notification ids.}
\usage{
UpdateNotificationIDs(old.ids, new.ids)
}
\arguments{
\item{old.ids}{A list of notificaion ids to be updated.}

\item{new.ids}{A list of new notification ids.}
}
\value{
A list of updated notification ids.
}
\description{
Update the existing list of notification ids.
}
\keyword{internal}
