% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_related_functions.R
\name{GenerateZipFileFromMutectVCFs}
\alias{GenerateZipFileFromMutectVCFs}
\title{This function generates a zip archive from Mutect VCF files.}
\usage{
GenerateZipFileFromMutectVCFs(
  files,
  zipfile,
  ref.genome,
  trans.ranges = NULL,
  region = "unknown",
  names.of.VCFs = NULL,
  tumor.col.names = NA,
  base.filename = "",
  updateProgress = NULL
)
}
\arguments{
\item{files}{Character vector of file paths to the Mutect VCF files.}

\item{zipfile}{Pathname of the zip file to be created.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link[ICAMS]{ICAMS}}.}

\item{trans.ranges}{Optional. If \code{ref.genome} specifies one of the
\code{\link{BSgenome}} object 
\enumerate{
  \item \code{\link[BSgenome.Hsapiens.1000genomes.hs37d5]{BSgenome.Hsapiens.1000genomes.hs37d5}}
  \item \code{\link[BSgenome.Hsapiens.UCSC.hg38]{BSgenome.Hsapiens.UCSC.hg38}}
  \item \code{\link[BSgenome.Mmusculus.UCSC.mm10]{BSgenome.Mmusculus.UCSC.mm10}}
}
then the function will infer \code{trans.ranges} automatically. Otherwise,
user will need to provide the necessary \code{trans.ranges}. Please refer to
\code{\link[ICAMS]{TranscriptRanges}} for more details.
If \code{is.null(trans.ranges)} do not add transcript range
information.}

\item{region}{A character string designating a genomic region;
see \code{\link[ICAMS]{as.catalog}} and \code{\link[ICAMS]{ICAMS}}.}

\item{names.of.VCFs}{Optional. Character vector of names of the VCF files.
The order of names in \code{names.of.VCFs} should match the order of VCFs
listed in \code{dir}. If \code{NULL}(default), this function will remove
all of the path up to and including the last path separator (if any) in
\code{dir} and file paths without extensions (and the leading dot) will be
used as the names of the VCF files.}

\item{tumor.col.names}{Optional. Character vector of column names in VCFs which contain
the tumor sample information. The order of names in \code{tumor.col.names}
should match the order of VCFs listed in \code{dir}. If
\code{tumor.col.names} is equal to \code{NA}(default), this function will
use the 10th column in all the VCFs to calculate VAFs.
See \code{\link[ICAMS]{GetMutectVAF}} for more details.}

\item{base.filename}{Optional. The base name of the CSV and PDF files to be
produced; multiple files will be generated, each ending in
\eqn{x}\code{.csv} or \eqn{x}\code{.pdf}, where \eqn{x} indicates the type
of catalog.}

\item{updateProgress}{Optional. Currently only used in ICAMS.shiny package to
update the progress indicator.}
}
\description{
This function generates a zip archive from Mutect VCF files.
}
\keyword{internal}
