% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_attribution.R
\name{GetExposureWithConfidence}
\alias{GetExposureWithConfidence}
\title{Get signature exposure for one sample with confidence interval}
\usage{
GetExposureWithConfidence(
  catalog,
  sig.universe,
  num.of.bootstrap.replicates = 1000,
  method = decomposeQP,
  conf.int = 0.95
)
}
\arguments{
\item{catalog}{A \strong{counts} catalog as defined in \code{\link{ICAMS}}.
It can only has \strong{one} column.}

\item{sig.universe}{The universe of signatures used to do signature attribution.}

\item{num.of.bootstrap.replicates}{The number of bootstrap replicates.}

\item{method}{Method used to get the optimal solution for signature attribution.}

\item{conf.int}{A number specifying the required confidence interval.}
}
\description{
Get signature exposure for one sample with confidence interval
}
\section{Value}{

A matrix showing the signature exposure results for
  \code{catalog} with lower and upper bound of the specified confidence
  interval.
}

